Dockerize project to build, test and run the game
Create a [Dockerfile][1] with [multi-stage building][2] to create
a build environment based on the [Gradle Docker image][3] and
a lightweight runner environment based on the [Eclipse Temurin
Docker image][4] (both based on OpenJDK 17).

Create a .dockerignore file based on .gitignore to skip files that
[should not be uploaded in the docker image][5].

Create a [docker-compose.yml][6] to setup four services:
- `game` to run the app for end users.
- `linter` to run all linters once.
- `tester` to rerun all tests with live reload.
- `demo` to rerun the app with live reload.

Add the same image name to [build the docker image only once][7] for
multiple services sharing with the same codebase.

Create a .env file to [configure environment variables][8] used inside
the compose file. Add the [`COMPOSE_PROJECT_NAME` variable][9] to assign
a prefix for all resources managed by compose.

Update the pre-commit config to use `docker compose` with the `linter`
image instead of raw `docker` commands to speed up the execution.

Update the README.md to use `docker compose` instead of raw `docker`
commands to execute tasks related to the project.

[1]: https://docs.docker.com/engine/reference/builder/
[2]: https://docs.docker.com/develop/develop-images/multistage-build/
[3]: https://hub.docker.com/_/gradle
[4]: https://hub.docker.com/_/eclipse-temurin
[5]: https://codefresh.io/docker-tutorial/not-ignore-dockerignore-2/
[6]: https://docs.docker.com/compose/compose-file/
[7]: https://stackoverflow.com/questions/40899236/how-to-prevent-docker-compose-building-the-same-image-multiple-times
[8]: https://docs.docker.com/compose/environment-variables/
[9]: https://docs.docker.com/compose/reference/envvars/#compose_project_name